backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "ollama>=0.6.1",
    "openai>=2.24.0",
    "python-dotenv>=1.2.2",
]
